Created variants of InfoViewTop to deal with helmet and capes having different data structures (arrays instead of objects).
